Switch int to long types

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        SalesManager manager = new SalesManager(new int[]{1, 2, 6});
+        SalesManager manager = new SalesManager(new long[]{1, 2, 6});
         System.out.println(manager.max());
     }
 }
